removed <vector> dependency moved some function declarations into cpp files

+++ addons/imguivariouscontrols/imguivariouscontrols.cpp
@@ -359,7 +359,7 @@ inline static bool ColorChooserInternal(ImVec4 *pColorOut,bool supportsAlpha,boo
             ImGui::ColorConvertHSVtoRGB(tint0, 1.0, 1.0, c0.x, c0.y, c0.z);
             ImGui::ColorConvertHSVtoRGB(tint1, 1.0, 1.0, c1.x, c1.y, c1.z);
 
-            window->DrawList->AddRectFilledMultiColor(window->Pos + pos, window->Pos + pos + ImVec2(width, quadSize / step),
+            window->DrawList->AddRectFilledMultiColor(window->Pos + pos, window->Pos + pos + ImVec2((float)width, quadSize / step),
                                                       ColorConvertFloat4ToU32(c0),
                                                       ColorConvertFloat4ToU32(c0),
                                                       ColorConvertFloat4ToU32(c1),
@@ -4288,7 +4288,7 @@ void LoadingIndicatorCircle2(const char* label,float indicatorRadiusFactor, floa
 
     int num_segments = 30;
 
-    int start = (int)fabs(ImSin(g.Time*1.8f)*(num_segments-5));
+    int start = (int)fabs(ImSin((float)(g.Time*1.8))*(num_segments-5));
 
     const float a_min = IM_PI*2.0f * ((float)start) / (float)num_segments;
     const float a_max = IM_PI*2.0f * ((float)num_segments-3) / (float)num_segments;
